Added ability to add enemies to levels

+++ src/Enemy.cpp
@@ -4,7 +4,7 @@
 Enemy::Enemy(SDL_Renderer* renderer) :
 	Bitmap(renderer, "Spikes.bmp", true)
 {
-	setSize(Vector2f(surface->w, surface->h));
+	setSize(Vector2f((float)surface->w, (float)surface->h));
 }
 
 void Enemy::playerCollision(std::shared_ptr<Level> level)
